Refactor : 불필요한 인터페이스 제거
서비스 클래스의 인터페이스 삭제
