+++ src/main/java/kr/codesqaud/cafe/repository/SignUpService.java
@@ -1,14 +1,31 @@
 package kr.codesqaud.cafe.repository;
 
 import kr.codesqaud.cafe.user.User;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface SignUpService {
-    void join(User user);
+@Service
+public class SignUpService {
 
-    Optional<User> findById(String userId);
+    @Autowired
+    private final MemberRepository memberRepository;
 
-    List<User> findAll();
+    public SignUpService(MemberRepository memberRepository) {
+        this.memberRepository = memberRepository;
+    }
+
+    public void join(User user) {
+        memberRepository.save(user);
+    }
+
+    public List<User> findAll() {
+        return memberRepository.findAll();
+    }
+
+    public Optional<User> findById(String userId) {
+        return memberRepository.findById(userId);
+    }
 }
