+++ src/test/java/kr/codesqaud/cafe/domain/repository/UserSaveTest.java
@@ -2,19 +2,16 @@ package kr.codesqaud.cafe.domain.repository;
 
 import kr.codesqaud.cafe.repository.MemoryMemberRepository;
 import kr.codesqaud.cafe.repository.SignUpService;
-import kr.codesqaud.cafe.repository.SignUpServiceImpl;
 import kr.codesqaud.cafe.user.User;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
-import java.util.Optional;
 
 class UserSaveTest {
 
-    private SignUpService signUpService = new SignUpServiceImpl(new MemoryMemberRepository());
+    private SignUpService signUpService = new SignUpService(new MemoryMemberRepository());
 
     @Test
     @DisplayName("회원 가입 잘 되는지 테스트")
